Refactored Create component
