+++ app/javascript/src/components/Dashboard/Notes/constants.js
@@ -1,0 +1,6 @@
+export const NOTE_FORM_INITIAL_VALUES = {
+  title: "",
+  description: "",
+  assigned: {},
+  tags: [],
+};
